+++ CHANGELOG.md
@@ -11,3 +11,13 @@ All notable changes to this repository will be documented in this file.
 
 - Restructure folder structure
 - added poetry dev script
+
+## [1.0.0] Sun, Feb 15, 2026
+
+- Added In-Memory cache strategy
+- DB dependency optional
+- Change UI
+
+## [1.0.3] Wed, Feb 18, 2026
+
+- Added Database connection retry logic